Enhance code (#12)
* Formatted code and added alert dialog for any unsaved changes

* Formatted code and added alert dialog for any unsaved changes

* Formatted code and added alert dialog for any unsaved changes

* removed warnings

* Optimized code by adding utility methods and removing redundant variable creation

* Optimized code by adding utility methods and removing redundant variable creation

* Changed image note layout, added scrim drawable and some ui fixed

* improved image note design

* removed unused code

* fixed search fragment crash and retaining query on configuration change

* improved look and feel for night mode and animation changes

* fixed fab button to slide up when snack bar appears

* improved theme of the app and some minor bug fixes

* added some new animations, changed theme colors, implemented floating edit text, removed unused strings

* formatted code and fixed some animations

* improved image note layout, added cardview to note detail layout, theme improvement and some enhancement in UI and code

* added image note editing functionality

+++ app/src/main/java/com/intkhabahmed/smartnotes/AddImageNote.java
@@ -5,8 +5,10 @@ import android.content.ContentValues;
 import android.content.Intent;
 import android.content.pm.PackageManager;
 import android.content.res.Resources;
+import android.database.Cursor;
 import android.graphics.Bitmap;
 import android.net.Uri;
+import android.os.Bundle;
 import android.preference.PreferenceManager;
 import android.provider.MediaStore;
 import android.support.annotation.NonNull;
@@ -15,9 +17,10 @@ import android.support.v4.content.ContextCompat;
 import android.support.v4.content.FileProvider;
 import android.support.v7.app.ActionBar;
 import android.support.v7.app.AppCompatActivity;
-import android.os.Bundle;
 import android.support.v7.widget.Toolbar;
+import android.text.Editable;
 import android.text.TextUtils;
+import android.text.TextWatcher;
 import android.view.Menu;
 import android.view.MenuItem;
 import android.view.View;
@@ -27,8 +30,10 @@ import android.widget.ImageButton;
 import android.widget.ImageView;
 import android.widget.Toast;
 
+import com.bumptech.glide.Glide;
 import com.intkhabahmed.smartnotes.notesdata.NotesContract;
 import com.intkhabahmed.smartnotes.utils.BitmapUtils;
+import com.intkhabahmed.smartnotes.utils.ViewUtils;
 
 import java.io.File;
 import java.io.IOException;
@@ -44,6 +49,11 @@ public class AddImageNote extends AppCompatActivity {
     private String mBackupTempImagePath;
     private Bitmap mResultBitmap;
     private Button mChangeImageButton;
+    private boolean mIsChanged;
+    private long mNoteId;
+    private boolean mIsEditing;
+    private boolean mIsImageChanged;
+    private String mOldDescription;
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
@@ -53,19 +63,39 @@ public class AddImageNote extends AppCompatActivity {
         Toolbar toolbar = findViewById(R.id.toolbar);
         setSupportActionBar(toolbar);
         ActionBar actionBar = getSupportActionBar();
-        if(actionBar != null) {
+        if (actionBar != null) {
             actionBar.setDisplayHomeAsUpEnabled(true);
             actionBar.setHomeAsUpIndicator(R.drawable.ic_arrow_white_black_24dp);
+            actionBar.setTitle(R.string.image_note);
         }
+
+        TextWatcher textWatcher = new TextWatcher() {
+            @Override
+            public void beforeTextChanged(CharSequence charSequence, int i, int i1, int i2) {
+
+            }
+
+            @Override
+            public void onTextChanged(CharSequence charSequence, int i, int i1, int i2) {
+                mIsChanged = !TextUtils.isEmpty(mNoteTitleEditText.getText().toString().trim());
+            }
+
+            @Override
+            public void afterTextChanged(Editable editable) {
+
+            }
+        };
 
         mCaptureImageButton = findViewById(R.id.capture_image_button);
         mNoteTitleEditText = findViewById(R.id.note_title_input);
         mImageView = findViewById(R.id.iv_image_note);
         mChangeImageButton = findViewById(R.id.change_image_button);
+        mNoteTitleEditText.addTextChangedListener(textWatcher);
 
         mChangeImageButton.setOnClickListener(new View.OnClickListener() {
             @Override
             public void onClick(View view) {
+                mBackupTempImagePath = mTempImagePath;
                 checkCameraPermission();
             }
         });
@@ -76,16 +106,43 @@ public class AddImageNote extends AppCompatActivity {
                 checkCameraPermission();
             }
         });
-        mImageView.setVisibility(View.GONE);
-        mChangeImageButton.setVisibility(View.INVISIBLE);
+        Intent intent = getIntent();
+        if (intent.hasExtra(Intent.EXTRA_TEXT)) {
+            mIsEditing = true;
+            mNoteId = intent.getLongExtra(Intent.EXTRA_TEXT, 0);
+            if (mNoteId > 0) {
+                Cursor cursor = getContentResolver().query(NotesContract.NotesEntry.CONTENT_URI,
+                        new String[]{NotesContract.NotesEntry.COLUMN_TITLE, NotesContract.NotesEntry.COLUMN_DESCRIPTION},
+                        NotesContract.NotesEntry._ID + "=?", new String[]{String.valueOf(mNoteId)}, null);
+                if (cursor != null) {
+                    cursor.moveToFirst();
+                    String title = cursor.getString(cursor.getColumnIndex(NotesContract.NotesEntry.COLUMN_TITLE));
+                    String description = cursor.getString(cursor.getColumnIndex(NotesContract.NotesEntry.COLUMN_DESCRIPTION));
+                    mNoteTitleEditText.setText(title);
+                    mOldDescription = description;
+                    File image = new File(description);
+                    if (image.exists()) {
+                        Glide.with(this).asDrawable().load(Uri.fromFile(image)).into(mImageView);
+                    }
+                    cursor.close();
+                }
+            }
+            mChangeImageButton.setVisibility(View.VISIBLE);
+            mCaptureImageButton.setVisibility(View.GONE);
+            mImageView.setVisibility(View.VISIBLE);
+        } else {
+            mIsEditing = false;
+            mImageView.setVisibility(View.GONE);
+            mChangeImageButton.setVisibility(View.INVISIBLE);
+        }
     }
 
     @Override
     public Resources.Theme getTheme() {
-        Resources.Theme theme =  super.getTheme();
+        Resources.Theme theme = super.getTheme();
         boolean isDarkThemeEnabled = PreferenceManager.getDefaultSharedPreferences(this)
                 .getBoolean(getString(R.string.dark_theme_key), false);
-        if(isDarkThemeEnabled){
+        if (isDarkThemeEnabled) {
             theme.applyStyle(R.style.AppThemeDark, true);
         } else {
             theme.applyStyle(R.style.AppThemeLight, true);
@@ -94,8 +151,8 @@ public class AddImageNote extends AppCompatActivity {
     }
 
     private void checkCameraPermission() {
-        if(ContextCompat.checkSelfPermission(this, Manifest.permission.WRITE_EXTERNAL_STORAGE) !=
-                PackageManager.PERMISSION_GRANTED){
+        if (ContextCompat.checkSelfPermission(this, Manifest.permission.WRITE_EXTERNAL_STORAGE) !=
+                PackageManager.PERMISSION_GRANTED) {
             ActivityCompat.requestPermissions(this, new String[]{Manifest.permission.WRITE_EXTERNAL_STORAGE},
                     RC_STORAGE_PERMISSION);
         } else {
@@ -107,7 +164,7 @@ public class AddImageNote extends AppCompatActivity {
     public void onRequestPermissionsResult(int requestCode, @NonNull String[] permissions, @NonNull int[] grantResults) {
         switch (requestCode) {
             case RC_STORAGE_PERMISSION:
-                if(grantResults.length > 0 && grantResults[0] == PackageManager.PERMISSION_GRANTED){
+                if (grantResults.length > 0 && grantResults[0] == PackageManager.PERMISSION_GRANTED) {
                     launchCamera();
                 } else {
                     Toast.makeText(this, "Storage Permission Denied", Toast.LENGTH_LONG).show();
@@ -117,15 +174,15 @@ public class AddImageNote extends AppCompatActivity {
 
     private void launchCamera() {
         Intent cameraIntent = new Intent(MediaStore.ACTION_IMAGE_CAPTURE);
-        if(cameraIntent.resolveActivity(getPackageManager()) != null){
+        if (cameraIntent.resolveActivity(getPackageManager()) != null) {
             File photo = null;
-            try{
+            try {
                 photo = BitmapUtils.createTempImageFile(this);
-            } catch (IOException ioe){
+            } catch (IOException ioe) {
                 ioe.printStackTrace();
             }
 
-            if(photo != null){
+            if (photo != null) {
                 mTempImagePath = photo.getAbsolutePath();
                 Uri photoUri = FileProvider.getUriForFile(this, FILEPROVIDER_AUTHORITY, photo);
                 cameraIntent.putExtra(MediaStore.EXTRA_OUTPUT, photoUri);
@@ -136,10 +193,13 @@ public class AddImageNote extends AppCompatActivity {
 
     @Override
     protected void onActivityResult(int requestCode, int resultCode, Intent data) {
-        if(requestCode == RC_CAPTURE_IMAGE && resultCode == RESULT_OK){
+        if (requestCode == RC_CAPTURE_IMAGE && resultCode == RESULT_OK) {
             processAndSetImage();
+            mIsChanged = true;
         } else {
-            BitmapUtils.deleteImageFile(this, mTempImagePath);
+            if (new File(mTempImagePath).exists()) {
+                BitmapUtils.deleteImageFile(this, mTempImagePath);
+            }
         }
         super.onActivityResult(requestCode, resultCode, data);
     }
@@ -148,27 +208,24 @@ public class AddImageNote extends AppCompatActivity {
         mImageView.setVisibility(View.VISIBLE);
         mCaptureImageButton.setVisibility(View.GONE);
         mResultBitmap = BitmapUtils.resamplePic(this, mTempImagePath);
-        if(mBackupTempImagePath != null){
+        if (mBackupTempImagePath != null && new File(mBackupTempImagePath).exists()) {
             BitmapUtils.deleteImageFile(AddImageNote.this, mBackupTempImagePath);
         }
         mImageView.setImageBitmap(mResultBitmap);
-        mImageView.setOnClickListener(new View.OnClickListener() {
-            @Override
-            public void onClick(View view) {
-                Toast.makeText(AddImageNote.this, "Image clicked", Toast.LENGTH_LONG).show();
-            }
-        });
         mChangeImageButton.setVisibility(View.VISIBLE);
     }
 
     @Override
     protected void onStop() {
         super.onStop();
-        if(mTempImagePath != null){
+        if (mTempImagePath != null && new File(mTempImagePath).exists()) {
             BitmapUtils.deleteImageFile(this, mTempImagePath);
         }
-        if(mBackupTempImagePath != null && (mTempImagePath != null && !mTempImagePath.equals(mBackupTempImagePath)) ){
+        if (mBackupTempImagePath != null && new File(mBackupTempImagePath).exists()) {
             BitmapUtils.deleteImageFile(this, mBackupTempImagePath);
+        }
+        if(mIsImageChanged) {
+            BitmapUtils.deleteImageFile(this, mOldDescription);
         }
     }
 
@@ -191,30 +248,52 @@ public class AddImageNote extends AppCompatActivity {
         return super.onOptionsItemSelected(item);
     }
 
+    @Override
+    public void onBackPressed() {
+        if (mIsChanged) {
+            ViewUtils.showUnsavedChangesDialog(this);
+        } else {
+            super.onBackPressed();
+        }
+        overridePendingTransition(R.anim.slide_in_left, R.anim.slide_out_right);
+    }
+
     private void insertImageNote() {
         String noteTitle = mNoteTitleEditText.getText().toString().trim();
-        if(!noteTitle.matches("[A-Za-z0-9]+") || noteTitle.matches("[0-9]+")){
-            Toast.makeText(this, "Title can only contain characters and numbers", Toast.LENGTH_LONG).show();
+        if (!noteTitle.matches("[A-Za-z0-9]+") || noteTitle.matches("[0-9]+")) {
+            Toast.makeText(this, "Title can only contain characters or characters and numbers", Toast.LENGTH_LONG).show();
             return;
         }
-        if(TextUtils.isEmpty(noteTitle)){
+        if (TextUtils.isEmpty(noteTitle)) {
             Toast.makeText(this, "Please enter a title of your note", Toast.LENGTH_LONG).show();
             return;
         }
-        if(mResultBitmap == null){
+        if (mResultBitmap == null) {
             Toast.makeText(this, "Please select an image for your note", Toast.LENGTH_LONG).show();
             return;
         }
         ContentValues values = new ContentValues();
         values.put(NotesContract.NotesEntry.COLUMN_TITLE, noteTitle);
         values.put(NotesContract.NotesEntry.COLUMN_DESCRIPTION, saveImageToStorage());
-        values.put(NotesContract.NotesEntry.COLUMN_TYPE, getString(R.string.image_note));
-        values.put(NotesContract.NotesEntry.COLUMN_DATE_CREATED, System.currentTimeMillis());
+        if (!mIsEditing) {
+            values.put(NotesContract.NotesEntry.COLUMN_TYPE, getString(R.string.image_note));
+            values.put(NotesContract.NotesEntry.COLUMN_DATE_CREATED, System.currentTimeMillis());
+            Uri uri = getContentResolver().insert(NotesContract.NotesEntry.CONTENT_URI, values);
+            if (uri != null) {
+                Toast.makeText(this, "Note created successfully!", Toast.LENGTH_LONG).show();
+                finish();
+                overridePendingTransition(R.anim.slide_in_left, R.anim.slide_out_right);
+                return;
+            }
+        }
+        mIsImageChanged = true;
         values.put(NotesContract.NotesEntry.COLUMN_DATE_MODIFIED, System.currentTimeMillis());
-        Uri uri = getContentResolver().insert(NotesContract.NotesEntry.CONTENT_URI, values);
-        if(uri != null){
-            Toast.makeText(this, "Note created successfully!", Toast.LENGTH_LONG).show();
+        int rowsUpdated = getContentResolver().update(NotesContract.NotesEntry.CONTENT_URI, values,
+                NotesContract.NotesEntry._ID + "=?", new String[]{String.valueOf(mNoteId)});
+        if (rowsUpdated > 0) {
+            Toast.makeText(this, "Note updated successfully!", Toast.LENGTH_LONG).show();
             finish();
+            overridePendingTransition(R.anim.slide_in_left, R.anim.slide_out_right);
         }
     }
